backend/requirements.txt
django==4.2.4 
celery==5.3.1 
redis==5.0.0 
psycopg2==2.9.7
python-dotenv==1.0.0
djangorestframework==3.14.0 
django-cors-headers==4.2.0